requirements.txt
appdirs==1.4.2
blinker==1.4
Django==1.10.5
django-extensions==1.7.7
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.5.4
mongoengine==0.11.0
packaging==16.8
pymongo==3.4.0
pyparsing==2.1.10
six==1.10.0
